requirements.txt
numpy>=1.24
scikit-learn>=1.3
pillow>=10.0
